+++ crates/amaru-protocols/src/blockfetch/responder.rs
@@ -15,7 +15,7 @@
 use std::fmt::Debug;
 
 use amaru_kernel::{IsHeader, NonEmptyVec, Point, RawBlock};
-use amaru_metrics::protocol::ServedBlockCountMetrics;
+use amaru_metrics::protocol::{ServedBlockCountMetrics, ServedBlockLatestMetrics};
 use amaru_observability::debug_span;
 use amaru_pure_stage::{DeserializerGuards, Effects, StageRef, Void};
 use tracing::Instrument;
@@ -169,8 +169,10 @@ impl StageState<State, Responder> for BlockFetchResponder {
         match input {
             StreamBlocks::Done => Ok((Some(ResponderAction::BatchDone), self)),
             StreamBlocks::More(points_range) => {
+                let point = *points_range.0.last();
                 let (block, points_range) = points_range.next_block(&store).await?;
                 Metrics::new(eff).record(ServedBlockCountMetrics { count: 1 }.into()).await;
+                Metrics::new(eff).record(ServedBlockLatestMetrics { slot: point.slot_or_default() }.into()).await;
                 // recurse if there are more blocks to fetch or signal that streaming is done
                 if let Some(points_range) = points_range {
                     eff.send(eff.me_ref(), Inputs::Local(StreamBlocks::More(points_range))).await;